refactor(playground): move Share state ownership to PlaygroundApp
`shareCopied` state and `handleShare` (clipboard write + 2s timeout) move
from NotationInput into PlaygroundApp, which is the single source of state
per the playground architecture rule. NotationInput receives `onShare` and
`shareCopied` as controlled props; Share button rendering stays in place.
